fis: move to execution utils
